AlgaeCal Test Submission
The cascading style sheet for the replica of the top of the AlgaeCal Product page: The positions are not precise as I did not apply all the margins and font sizes from 'view-source'. I tried to "eye" the positions as best as I could without copying from the 'view-source' page. The Glober font is in the 'font' folder.

The HTML site for the replica of the top of the AlgaeCal Product page: The images were found from 'view-source' and are all in the 'images' folder and there is a image placeholder for the wp-plugin.

The Carousel post type plugin: The plugin is incomplete. I created a custom post type (carousel_cpt.php) page  and an admin page from scratch  (carousel_main.php & cs_carousel_admin1.php, mycss.ss, and myjs.js ) to attempt to work around the issues I ran into.

Lastly, christinshelton_wordpress_script.php is included to show the work of Part Two and is "included" in carousel_cpt.php